src/python/tpl-gen/pyproject.toml
[tool.poetry]
name = "tpl-gen"
version = "1.7.0"
description = "A template generator using Jinja2 and Smarty TPL files."
authors = ["Engineering <engineering@csitea.net>"]
maintainers = ["Engineering <engineering@csitea.net>"]

packages = [
    { include = "tpl_gen" },
    { include = "tpl_gen/libs"},
    { include = "tpl_gen/config"},

]

[tool.poetry.dependencies]
python = ">3.8,<4"
pprintjson = "^1.4.2"
jinja2 = "^3.0.1"
colorama = "^0.4.4"
pyyaml = "^5.3.1"
botocore = "^1.29.152"
boto3-type-annotations = "^0.3.1"
boto3 = "^1.26.152"
requests = "^2.28.1"
rich = "^13.3.6"
jq = "^1.4.1"
setuptools="^67.8.0"


[tool.poetry.dev-dependencies]
pylint = "^2.8.3"
autopep8 = "^1.5.7"
pytest = "^6.2.4"
pytest-mock = "^3.6.1"
coverage = "^5.5"
pyyaml = "^5.3.1"

[tool.poetry.scripts]
start  = "tpl_gen:main"
render = "tpl_gen.convert_yaml:yaml_to_json"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
